langchain/chatbot/requirements.txt
mypy
ruff
pytest
python-dotenv
langchain
langchain-openai
langgraph-cli[inmem]